Add MS check for HR logic

+++ includes/class-wp-print-preview-helper.php
@@ -2,6 +2,18 @@
 
 class Wp_Print_Preview_Helper
 {
+	// CREATED FROM EXCHANGE TABLE
+	const HR_DEPARTMENTS = [
+		'ACRE',
+		'Benefits',
+		'Human Resources',
+		'Labor Relations',
+		'Retirement Services',
+		'Selection & Class',
+		'Substance Abuse',
+		'WD/EX'
+	];
+
     /**
      * Connected to WC hook callback "bc_entry_id_text_to_order_items()"
      * @param $entry - Gravity Forms entry object
@@ -185,6 +197,9 @@ class Wp_Print_Preview_Helper
             array_push( $text_params_arr, $fax_label_text, $fax_text );
         }
 
+		// CHECK IF USER IS HR
+	    $is_HR = $this->_checkIfUserHR((string)$email);
+
         $image = new \Imagick();
         $image->readImage( plugin_dir_path( __FILE__ ) . '../public/assets/blank.png' );
         $image->setImageColorspace( Imagick::COLORSPACE_SRGB );
@@ -251,17 +266,17 @@ class Wp_Print_Preview_Helper
 	        // use magick if not in localhost
 	        $magick_cmd = is_int(strpos(site_url(), 'local')) ? '' : 'magick';
 			// convert to PDF & attach HR backing
-	        // TODO - logic for HR backing
-            exec( "$magick_cmd convert $source -resize 50% $target && " .
-            "$magick_cmd convert -density 600 $target $hr_backing $target 2>&1", $output, $res );
-
+	        if ( $is_HR ) {
+		        exec("$magick_cmd convert $source -resize 50% $target && " .
+		             "$magick_cmd convert -density 600 $target $hr_backing $target 2>&1", $output, $res);
+	        }
             if ( $res > 0 ) {
                 error_log( json_encode( $output, JSON_PRETTY_PRINT ) );
             }
 
             // create 25 up output on 12 x 18 if flag is set
             if ( $create25up ) {
-                $this->_create25Up( $image, $entry_filename );
+                $this->_create25Up( $image, $entry_filename, $is_HR );
             }
         }
 
@@ -404,8 +419,9 @@ class Wp_Print_Preview_Helper
      * Creates and writes a 25 up Business Card printout on a 12 x 18 stock
      * @param $image - Image Magick object (Business Card)
      * @param $filename - filename of business card PDF
+     * @param bool $is_HR - if this 25-up is for HR employee.
      */
-    private function _create25Up( $image, $filename )
+    private function _create25Up( $image, $filename, bool $is_HR )
     {
         // new filename for 25-up PDF
 //        $new_filename = $filename . '_25_up.pdf';
@@ -450,7 +466,7 @@ class Wp_Print_Preview_Helper
 
         // Create targets for temp 25-up PNG and final 25-up PDF output
 	    $temp_25_png  = $uploads_dir['basedir'] . '/business_cards/' . $filename . '_25_up.png';
-	    $temp_25_png_hr  = $uploads_dir['basedir'] . '/business_cards/' . $filename . '_25_up_HR.png';
+	    // $temp_25_png_hr  = $uploads_dir['basedir'] . '/business_cards/' . $filename . '_25_up_HR.png';
 	    $final_25_pdf = $uploads_dir['basedir'] . '/business_cards/' . $filename . '_25_up.pdf';
 	    $final_25_pdf_hr = $assets_dir . 'HR-buscardback-final-25.pdf';
 
@@ -464,22 +480,24 @@ class Wp_Print_Preview_Helper
          * 1. Create 25-up with PNG to retain resolution
          * 2. Convert to PDF and add border.
          * 3. Remove temp 25-up PNG from business_cards directory
-         * 4. Add HR 25-up backing to page 2.
+         * 4. Add HR 25-up backing to page 2 (if applicable).
          */
         // define PATH
         putenv( 'PATH=/usr/local/sbin:/usr/local/bin:/usr/sbin:/usr/bin:/sbin:/bin' );
 		// use magick if not in localhost
 		$magick_cmd = is_int(strpos(site_url(), 'local')) ? '' : 'magick';
 
+		$sh_cmd = "($magick_cmd montage " . $base_png . ' -mode concatenate -duplicate 24 -tile 5x5 ' . $temp_25_png . ' && ' .
+		          "$magick_cmd convert " . $temp_25_png . ' -density 600 -bordercolor white -border 150x600 ' . $final_25_pdf ;
+
+		// add HR backing if applicable.
+		if ( $is_HR ) {
+			$sh_cmd .= " && $magick_cmd convert -density 600 $final_25_pdf $final_25_pdf_hr $final_25_pdf &&" .
+			           'rm ' . $temp_25_png . ' 2>&1) > /dev/null 2>/dev/null &';
+		}
+
 	    // TODO - add logic for HR backing
-        exec(
-            "($magick_cmd montage " . $base_png . ' -mode concatenate -duplicate 24 -tile 5x5 ' . $temp_25_png . ' && ' .
-            "$magick_cmd convert " . $temp_25_png . ' -density 600 -bordercolor white -border 150x600 ' . $final_25_pdf . ' && ' .
-            "$magick_cmd convert -density 600 $final_25_pdf $final_25_pdf_hr $final_25_pdf &&" .
-            'rm ' . $temp_25_png . ' 2>&1) > /dev/null 2>/dev/null &',
-            $output,
-            $res
-        );
+        exec($sh_cmd, $output, $res);
 	    // exit with error code if command unsuccessful
 	    if ( $res > 0 ) {
 		    error_log( 'Error Code: ' . $res . "\n" . json_encode( $output, JSON_PRETTY_PRINT ) );
@@ -505,4 +523,33 @@ class Wp_Print_Preview_Helper
 		//     exit( $res_HR );
 	    // }
     }
+
+	/**
+	 * Checks if the provided email is a VTA user that's under the HR division.
+	 * Should match one of the HR departments.
+	 * @param string $email email or UPN.
+	 * @return bool if user is in HR.
+	 */
+	private function _checkIfUserHR( string $email ): bool {
+		try {
+			$graph_api = new wp_print_preview\utils\MsGraphApi();
+
+			// try email first, then UPN
+			$user = $graph_api->search_user_by_email($email);
+			if (!$user)
+				$user = $graph_api->search_user_by_upn($email);
+			if (!$user)
+				// if user still isn't found, then default to false.
+				return false;
+
+			foreach ( self::HR_DEPARTMENTS as $department )
+				if ( preg_match("/$department/", $user->getDepartment(), $matches, 'i') )
+					return true;
+			return false;
+
+		} catch ( Exception $e ) {
+			error_log("Wp_Print_Preview_Helper::_checkIfUserHR() error - $e");
+			return false;
+		}
+	}
 }
